Move common RobotProperty structs to rmf_site_format
