+++ rmf_site_editor/src/widgets/inspector/inspect_model_description/inspect_mobility.rs
@@ -18,72 +18,21 @@
 use super::{
     get_selected_description_entity,
     inspect_robot_properties::{
-        serialize_and_change_robot_property, show_robot_property_widget, RecallPropertyKind,
-        RobotProperty, RobotPropertyKind, RobotPropertyWidgetRegistry,
+        serialize_and_change_robot_property, show_robot_property_widget,
+        RobotPropertyWidgetRegistry,
     },
     ModelPropertyQuery,
 };
 use crate::{
-    site::{Change, Group, ModelMarker, ModelProperty, Robot},
+    site::{
+        Change, DifferentialDrive, Group, Mobility, ModelMarker, ModelProperty, RecallMobility,
+        Robot, RobotProperty,
+    },
     widgets::{prelude::*, Inspect},
 };
 use bevy::{ecs::system::SystemParam, prelude::*};
 use bevy_egui::egui::{DragValue, Grid, Ui};
-use rmf_site_format::Recall;
-use serde::{Deserialize, Serialize};
-use serde_json::Map;
 use smallvec::SmallVec;
-
-#[derive(Serialize, Deserialize, Debug, Clone, Component, PartialEq)]
-pub struct Mobility {
-    pub kind: String,
-    pub config: serde_json::Value,
-}
-
-impl Default for Mobility {
-    fn default() -> Self {
-        Self {
-            kind: String::new(),
-            config: serde_json::Value::Object(Map::new()),
-        }
-    }
-}
-
-impl RobotProperty for Mobility {
-    fn new(kind: String, config: serde_json::Value) -> Self {
-        Self { kind, config }
-    }
-
-    fn is_default(&self) -> bool {
-        if *self == Self::default() {
-            return true;
-        }
-        false
-    }
-
-    fn kind(&self) -> Option<String> {
-        Some(self.kind.clone())
-    }
-
-    fn label() -> String {
-        "Mobility".to_string()
-    }
-}
-
-#[derive(Clone, Debug, Default, Component, PartialEq)]
-pub struct RecallMobility {
-    pub kind: Option<String>,
-    pub config: Option<serde_json::Value>,
-}
-
-impl Recall for RecallMobility {
-    type Source = Mobility;
-
-    fn remember(&mut self, source: &Mobility) {
-        self.kind = Some(source.kind.clone());
-        self.config = Some(source.config.clone());
-    }
-}
 
 #[derive(SystemParam)]
 pub struct InspectMobility<'w, 's> {
@@ -159,75 +108,6 @@ impl<'w, 's> WidgetSystem<Inspect> for InspectMobility<'w, 's> {
                 let _ = world.try_show_in(child, inspect, ui);
             }
         }
-    }
-}
-
-// Supported kinds of Mobility
-#[derive(Serialize, Deserialize, Debug, Clone, Component, PartialEq, Reflect)]
-#[reflect(Component)]
-pub struct DifferentialDrive {
-    pub bidirectional: bool,
-    pub rotation_center_offset: [f32; 2],
-    pub translational_speed: f32,
-    pub translational_acceleration: f32,
-    pub rotational_speed: f32,
-    pub rotational_acceleration: f32,
-}
-
-impl Default for DifferentialDrive {
-    fn default() -> Self {
-        Self {
-            bidirectional: false,
-            rotation_center_offset: [0.0, 0.0],
-            translational_speed: 0.5,
-            translational_acceleration: 0.25,
-            rotational_speed: 1.0,
-            rotational_acceleration: 1.5,
-        }
-    }
-}
-
-impl RobotPropertyKind for DifferentialDrive {
-    fn label() -> String {
-        "Differential Drive".to_string()
-    }
-}
-
-#[derive(Clone, Debug, Default, Component, PartialEq)]
-pub struct RecallDifferentialDrive {
-    pub bidirectional: Option<bool>,
-    pub rotation_center_offset: Option<[f32; 2]>,
-    pub translational_speed: Option<f32>,
-    pub translational_acceleration: Option<f32>,
-    pub rotational_speed: Option<f32>,
-    pub rotational_acceleration: Option<f32>,
-}
-
-impl RecallPropertyKind for RecallDifferentialDrive {
-    type Kind = DifferentialDrive;
-
-    fn assume(&self) -> DifferentialDrive {
-        DifferentialDrive {
-            bidirectional: self.bidirectional.clone().unwrap_or_default(),
-            rotation_center_offset: self.rotation_center_offset.clone().unwrap_or_default(),
-            translational_speed: self.translational_speed.clone().unwrap_or_default(),
-            translational_acceleration: self.translational_acceleration.clone().unwrap_or_default(),
-            rotational_speed: self.rotational_speed.clone().unwrap_or_default(),
-            rotational_acceleration: self.rotational_acceleration.clone().unwrap_or_default(),
-        }
-    }
-}
-
-impl Recall for RecallDifferentialDrive {
-    type Source = DifferentialDrive;
-
-    fn remember(&mut self, source: &DifferentialDrive) {
-        self.bidirectional = Some(source.bidirectional);
-        self.rotation_center_offset = Some(source.rotation_center_offset);
-        self.translational_speed = Some(source.translational_speed);
-        self.translational_acceleration = Some(source.translational_acceleration);
-        self.rotational_speed = Some(source.rotational_speed);
-        self.rotational_acceleration = Some(source.rotational_acceleration);
     }
 }
